Profile tests

+++ accounts/views.py
@@ -1,34 +1,48 @@
 from rest_framework import mixins, status, viewsets
 from rest_framework.exceptions import PermissionDenied
-from rest_framework.permissions import IsAuthenticated
+from rest_framework.permissions import AllowAny, IsAuthenticated
 from rest_framework.response import Response
 
 from accounts.models import Profile
+from accounts.permissions import IsOwner
 
 from .serializers import ProfileNestedSerializer, ProfileSerializer
 
 
 class ProfileViewSet(
-    mixins.RetrieveModelMixin, mixins.UpdateModelMixin, mixins.CreateModelMixin, viewsets.GenericViewSet
+    mixins.RetrieveModelMixin,
+    mixins.ListModelMixin,
+    mixins.UpdateModelMixin,
+    mixins.CreateModelMixin,
+    viewsets.GenericViewSet,
 ):
     queryset = Profile.objects.all()
     serializer_class = ProfileSerializer
-    retrieve_serializer_class = ProfileNestedSerializer
+    retrieve_list_serializer_class = ProfileNestedSerializer
     permission_classes = [IsAuthenticated]
 
     def get_serializer_class(self):
-        if self.action == "retrieve":
-            if hasattr(self, "retrieve_serializer_class"):
-                return self.retrieve_serializer_class
+        if self.action == "retrieve" or self.action == "list":
+            if hasattr(self, "retrieve_list_serializer_class"):
+                return self.retrieve_list_serializer_class
         return self.serializer_class
+
+    def get_permissions(self):
+        if self.action == "retrieve" or self.action == "list":
+            permission_classes = [AllowAny]
+        elif self.action == "create":
+            permission_classes = [IsAuthenticated]
+        else:
+            permission_classes = [IsAuthenticated & IsOwner]
+        return [permission() for permission in permission_classes]
 
     def create(self, request, *args, **kwargs):
         if hasattr(request.user, "profile") == True:
             raise PermissionDenied({"detail": "User already has profile."})
 
-        request.data["owner"] = request.user.id
         serializer = self.get_serializer(data=request.data)
         serializer.is_valid(raise_exception=True)
+        serializer.validated_data["owner"] = request.user
         self.perform_create(serializer)
         return Response(serializer.data, status=status.HTTP_201_CREATED)
 
@@ -38,7 +52,8 @@ class ProfileViewSet(
         serializer = self.get_serializer(instance, data=request.data, partial=partial)
         serializer.is_valid(raise_exception=True)
         if "avatar" in request.data:
-            instance.avatar.delete()
+            if instance.avatar != "default/avatar.png":
+                instance.avatar.delete()
         self.perform_update(serializer)
 
         if getattr(instance, "_prefetched_objects_cache", None):
